+++ app/src/main/res/menu/bottom_nav.xml
@@ -14,7 +14,7 @@
         android:icon="@drawable/ic_sos_item_menu"
         android:title="@string/sos_bottom_menu_item" />
     <item
-        android:id="@+id/user"
+        android:id="@+id/LoginFragment"
         android:icon="@drawable/ic_user_item_menu"
         android:title="@string/user_bottom_menu_item" />
 </menu>
