+++ features/form/src/main/java/com/petsaverapp/form/ui/fragments/LoginFragment.kt
@@ -1,0 +1,29 @@
+package com.petsaverapp.form.ui.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.petsaverapp.form.databinding.FragmentLoginBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LoginFragment : Fragment() {
+    private var _binding: FragmentLoginBinding? = null
+    private val binding get() = _binding!!
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+       _binding = FragmentLoginBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+}
