+++ features/form/src/main/res/drawable/ic_login_email.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="23dp"
+    android:height="23dp"
+    android:viewportWidth="23"
+    android:viewportHeight="23">
+  <path
+      android:pathData="M22.944,4.957C22.848,4.474 22.632,4.018 22.318,3.634C22.254,3.553 22.187,3.482 22.114,3.409C21.555,2.847 20.777,2.524 19.979,2.524H3.021C2.214,2.524 1.456,2.838 0.885,3.409C0.813,3.481 0.746,3.554 0.68,3.637C0.367,4.019 0.152,4.475 0.059,4.959C0.02,5.149 0,5.346 0,5.545V17.455C0,17.87 0.084,18.273 0.251,18.657C0.395,18.997 0.615,19.32 0.885,19.59C0.953,19.658 1.021,19.721 1.093,19.782C1.634,20.229 2.318,20.476 3.021,20.476H19.979C20.686,20.476 21.37,20.229 21.909,19.777C21.981,19.719 22.048,19.657 22.115,19.59C22.376,19.329 22.584,19.029 22.733,18.699L22.753,18.652C22.917,18.275 23,17.873 23,17.455V5.545C23,5.348 22.981,5.15 22.944,4.957ZM1.564,4.452C1.607,4.389 1.662,4.325 1.731,4.255C2.077,3.91 2.535,3.72 3.021,3.72H19.979C20.469,3.72 20.928,3.91 21.269,4.256C21.328,4.315 21.385,4.381 21.434,4.448L21.563,4.622L12.512,12.511C12.233,12.755 11.873,12.89 11.5,12.89C11.13,12.89 10.771,12.756 10.489,12.511L1.446,4.624L1.564,4.452ZM1.202,17.56C1.197,17.528 1.196,17.492 1.196,17.455V5.791L8.22,11.919L1.267,17.982L1.202,17.56ZM20.884,19.039C20.612,19.196 20.299,19.28 19.979,19.28H3.021C2.701,19.28 2.388,19.196 2.116,19.039L1.832,18.875L9.016,12.614L9.804,13.299C10.276,13.708 10.878,13.935 11.5,13.935C12.124,13.935 12.727,13.708 13.199,13.299L13.986,12.613L21.168,18.875L20.884,19.039ZM21.804,17.455C21.804,17.491 21.803,17.527 21.799,17.559L21.736,17.984L14.78,11.922L21.804,5.794V17.455Z"
+      android:fillColor="#7F7F7F"
+      android:fillAlpha="0.5"/>
+  <path
+      android:pathData="M22.944,4.957C22.848,4.474 22.632,4.018 22.318,3.634C22.254,3.553 22.187,3.482 22.114,3.409C21.555,2.847 20.777,2.524 19.979,2.524H3.021C2.214,2.524 1.456,2.838 0.885,3.409C0.813,3.481 0.746,3.554 0.68,3.637C0.367,4.019 0.152,4.475 0.059,4.959C0.02,5.149 0,5.346 0,5.545V17.455C0,17.87 0.084,18.273 0.251,18.657C0.395,18.997 0.615,19.32 0.885,19.59C0.953,19.658 1.021,19.721 1.093,19.782C1.634,20.229 2.318,20.476 3.021,20.476H19.979C20.686,20.476 21.37,20.229 21.909,19.777C21.981,19.719 22.048,19.657 22.115,19.59C22.376,19.329 22.584,19.029 22.733,18.699L22.753,18.652C22.917,18.275 23,17.873 23,17.455V5.545C23,5.348 22.981,5.15 22.944,4.957ZM1.564,4.452C1.607,4.389 1.662,4.325 1.731,4.255C2.077,3.91 2.535,3.72 3.021,3.72H19.979C20.469,3.72 20.928,3.91 21.269,4.256C21.328,4.315 21.385,4.381 21.434,4.448L21.563,4.622L12.512,12.511C12.233,12.755 11.873,12.89 11.5,12.89C11.13,12.89 10.771,12.756 10.489,12.511L1.446,4.624L1.564,4.452ZM1.202,17.56C1.197,17.528 1.196,17.492 1.196,17.455V5.791L8.22,11.919L1.267,17.982L1.202,17.56ZM20.884,19.039C20.612,19.196 20.299,19.28 19.979,19.28H3.021C2.701,19.28 2.388,19.196 2.116,19.039L1.832,18.875L9.016,12.614L9.804,13.299C10.276,13.708 10.878,13.935 11.5,13.935C12.124,13.935 12.727,13.708 13.199,13.299L13.986,12.613L21.168,18.875L20.884,19.039ZM21.804,17.455C21.804,17.491 21.803,17.527 21.799,17.559L21.736,17.984L14.78,11.922L21.804,5.794V17.455Z"
+      android:fillColor="#3D3D3D"/>
+</vector>
