+++ features/form/src/main/res/drawable/ic_login_senha.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="23dp"
+    android:height="23dp"
+    android:viewportWidth="23"
+    android:viewportHeight="23">
+  <group>
+    <clip-path
+        android:pathData="M0,0h23v23h-23z"/>
+    <path
+        android:pathData="M20.724,8.865C20.014,8.152 19.018,7.707 17.932,7.708H17.53V6.029C17.531,4.369 16.853,2.853 15.763,1.766C14.676,0.677 13.161,-0.001 11.5,0C9.839,-0.001 8.323,0.677 7.236,1.766C6.147,2.853 5.469,4.369 5.47,6.029V7.708H5.067C3.981,7.707 2.986,8.152 2.276,8.865C1.563,9.575 1.118,10.571 1.119,11.657V19.051C1.118,20.138 1.563,21.133 2.276,21.843C2.986,22.556 3.981,23.001 5.067,23H17.932C19.018,23.001 20.014,22.556 20.724,21.843C21.437,21.133 21.882,20.138 21.881,19.051L21.881,15.41L21.881,11.657C21.882,10.571 21.437,9.575 20.724,8.865ZM8.081,6.029C8.082,5.082 8.462,4.235 9.083,3.612C9.705,2.992 10.552,2.611 11.5,2.611C12.448,2.611 13.295,2.992 13.917,3.612C14.539,4.235 14.918,5.082 14.919,6.029V7.708H8.081V6.029ZM18.878,19.997C18.632,20.242 18.306,20.389 17.932,20.389H5.067C4.694,20.389 4.368,20.242 4.122,19.997C3.877,19.752 3.731,19.424 3.73,19.051V11.657C3.731,11.283 3.877,10.957 4.122,10.711C4.368,10.467 4.694,10.319 5.067,10.319H17.932C18.306,10.319 18.632,10.467 18.878,10.711C19.122,10.957 19.269,11.283 19.27,11.657V19.051C19.269,19.424 19.122,19.752 18.878,19.997Z"
+        android:fillColor="#7F7F7F"
+        android:fillAlpha="0.5"/>
+    <path
+        android:pathData="M20.724,8.865C20.014,8.152 19.018,7.707 17.932,7.708H17.53V6.029C17.531,4.369 16.853,2.853 15.763,1.766C14.676,0.677 13.161,-0.001 11.5,0C9.839,-0.001 8.323,0.677 7.236,1.766C6.147,2.853 5.469,4.369 5.47,6.029V7.708H5.067C3.981,7.707 2.986,8.152 2.276,8.865C1.563,9.575 1.118,10.571 1.119,11.657V19.051C1.118,20.138 1.563,21.133 2.276,21.843C2.986,22.556 3.981,23.001 5.067,23H17.932C19.018,23.001 20.014,22.556 20.724,21.843C21.437,21.133 21.882,20.138 21.881,19.051L21.881,15.41L21.881,11.657C21.882,10.571 21.437,9.575 20.724,8.865ZM8.081,6.029C8.082,5.082 8.462,4.235 9.083,3.612C9.705,2.992 10.552,2.611 11.5,2.611C12.448,2.611 13.295,2.992 13.917,3.612C14.539,4.235 14.918,5.082 14.919,6.029V7.708H8.081V6.029ZM18.878,19.997C18.632,20.242 18.306,20.389 17.932,20.389H5.067C4.694,20.389 4.368,20.242 4.122,19.997C3.877,19.752 3.731,19.424 3.73,19.051V11.657C3.731,11.283 3.877,10.957 4.122,10.711C4.368,10.467 4.694,10.319 5.067,10.319H17.932C18.306,10.319 18.632,10.467 18.878,10.711C19.122,10.957 19.269,11.283 19.27,11.657V19.051C19.269,19.424 19.122,19.752 18.878,19.997Z"
+        android:fillColor="#3D3D3D"/>
+    <path
+        android:pathData="M11.499,12.577C10.461,12.577 9.62,13.419 9.62,14.458C9.62,15.223 10.079,15.878 10.735,16.172L10.319,18.871H11.499H12.681L12.264,16.172C12.921,15.878 13.38,15.223 13.38,14.458C13.38,13.419 12.538,12.577 11.499,12.577Z"
+        android:fillColor="#7F7F7F"
+        android:fillAlpha="0.5"/>
+    <path
+        android:pathData="M11.499,12.577C10.461,12.577 9.62,13.419 9.62,14.458C9.62,15.223 10.079,15.878 10.735,16.172L10.319,18.871H11.499H12.681L12.264,16.172C12.921,15.878 13.38,15.223 13.38,14.458C13.38,13.419 12.538,12.577 11.499,12.577Z"
+        android:fillColor="#3D3D3D"/>
+  </group>
+</vector>
