+++ features/form/src/main/res/layout/fragment_login.xml
@@ -1,20 +1,18 @@
-<?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue"
-    tools:context=".login">
+    tools:context=".ui.fragments.LoginFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/fundo_branco_login"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="20dp"
+        android:layout_marginTop="20dp"
         android:layout_marginEnd="20dp"
         android:background="@drawable/fundo_quadrado_bordas_arredondados"
-        app:layout_constraintBottom_toTopOf="@+id/guideline"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
@@ -26,10 +24,21 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
             android:layout_marginTop="20dp"
+            android:fontFamily="@font/inter_medium"
             android:text="@string/facaagoraseulogin"
             android:textColor="@color/mid_grey"
             app:layout_constraintBottom_toTopOf="@id/subtitulo_login"
             app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <Button
+            android:layout_width="20dp"
+            android:layout_height="20dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginRight="20dp"
+            android:background="@drawable/ic_login_sair"
+            android:drawableTint="@color/mid_grey"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
@@ -38,6 +47,7 @@
             android:layout_height="wrap_content"
             android:layout_marginStart="20dp"
             android:layout_marginTop="10dp"
+            android:fontFamily="@font/inter_extra_bold"
             android:text="@string/login"
             android:textColor="@color/dark_grey"
             android:textSize="28sp"
@@ -49,22 +59,29 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
+            android:fontFamily="@font/inter_medium"
             android:paddingStart="20dp"
             android:text="@string/subtitulo_login_3"
+            android:textSize="15sp"
             app:layout_constraintEnd_toStartOf="@+id/guideline2"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/subtitulo_login" />
 
         <EditText
+            android:id="@+id/editText_login_email"
             android:layout_width="match_parent"
-            android:layout_height="30dp"
+            android:layout_height="48dp"
             android:layout_marginStart="20dp"
             android:layout_marginTop="20dp"
             android:layout_marginEnd="20dp"
             android:background="@drawable/background_input_text"
             android:backgroundTint="@color/grey"
-            android:hint="Email"
-            android:drawableStart="@drawable/ic_home_item_menu"
+            android:drawableStart="@drawable/ic_login_email"
+            android:drawablePadding="10dp"
+            android:drawableTint="@color/mid_grey"
+            android:hint="@string/email"
+            android:inputType="textEmailAddress"
+            android:paddingLeft="10dp"
             android:textColorHint="@color/mid_grey"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -72,12 +89,78 @@
 
         </EditText>
 
+        <EditText
+            android:id="@+id/editeText_login_senha"
+            android:layout_width="match_parent"
+            android:layout_height="48dp"
+            android:layout_marginStart="20dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginEnd="20dp"
+            android:background="@drawable/background_input_text"
+            android:backgroundTint="@color/grey"
+            android:drawableStart="@drawable/ic_login_senha"
+            android:drawablePadding="10dp"
+            android:drawableTint="@color/mid_grey"
+            android:hint="@string/senha"
+            android:inputType="textPassword"
+            android:paddingLeft="10dp"
+            android:textColorHint="@color/mid_grey"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/editText_login_email"></EditText>
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/btn_bemVindo_home"
+            style="@style/Widget.Material3.Button.ElevatedButton.Icon"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginEnd="20dp"
+            android:background="@drawable/background_input_text"
+            android:backgroundTint="@color/blue"
+            android:fontFamily="@font/inter_medium"
+            android:text="@string/entrar"
+            android:textAlignment="center"
+            android:textColor="@color/white"
+            android:textFontWeight="8"
+            android:textSize="18sp"
+            app:backgroundTint="@color/dark_blue"
+            app:elevation="10dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/editeText_login_senha" />
+
+        <TextView
+            android:id="@+id/texto_informativo_tela_login"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginBottom="50dp"
+            android:text="@string/textoInformativoTelaLogin"
+            android:textColor="@color/mid_grey"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/btn_bemVindo_home" />
+
+        <TextView
+            android:id="@+id/btn_cadastrese_tela_login"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="4dp"
+            android:layout_marginTop="20dp"
+            android:text="@string/cadastrese"
+            android:textColor="@color/blue"
+            app:layout_constraintStart_toEndOf="@+id/texto_informativo_tela_login"
+            app:layout_constraintTop_toBottomOf="@+id/btn_bemVindo_home" />
+
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guideline2"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.65" />
+            app:layout_constraintGuide_percent="0.80" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <androidx.constraintlayout.widget.Guideline
