+++ features/form/src/main/res/values/strings.xml
@@ -19,4 +19,9 @@
     <string name="facaagoraseulogin">Faça agora seu login!</string>
     <string name="login">Login</string>
     <string name="subtitulo_login_3">Seja bem vindo de volta, digite seu email e senha para continuar!</string>
+    <string name="senha">Senha</string>
+    <string name="email">Email</string>
+    <string name="entrar">Entrar</string>
+    <string name="textoInformativoTelaLogin">Ainda não tem uma conta?</string>
+    <string name="cadastrese">Cadastre-se</string>
 </resources>
